requirements.txt
# Core dependencies for time series analysis
pandas>=1.5.0
numpy>=1.21.0
prophet>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly

# Additional dependencies for AI analysis
langchain-openai
python-dotenv